Rename files for convention
