+++ src/renderer/components/kube-object-menu/dependencies/api-manager.injectable.ts
