+++ src/renderer/components/kube-object-menu/dependencies/cluster-name.injectable.ts
