+++ src/renderer/components/kube-object-menu/dependencies/edit-resource-tab.injectable.ts
