+++ src/renderer/components/kube-object-menu/dependencies/hide-details.injectable.ts
