+++ src/renderer/components/kube-object-menu/kube-object-menu.test.tsx
@@ -32,14 +32,14 @@ import asyncFn, { AsyncFnMock } from "@async-fn/jest";
 import { getDiForUnitTesting } from "../getDiForUnitTesting";
 
 import clusterInjectable from "./dependencies/cluster.injectable";
-import hideDetailsInjectable from "./dependencies/hideDetails.injectable";
-import editResourceTabInjectable from "./dependencies/editResourceTab.injectable";
+import hideDetailsInjectable from "./dependencies/hide-details.injectable";
+import editResourceTabInjectable from "./dependencies/edit-resource-tab.injectable";
 import { TabKind } from "../dock/dock.store";
-import kubeObjectMenuRegistryInjectable from "./dependencies/kubeObjectMenuRegistry.injectable";
+import kubeObjectMenuRegistryInjectable from "./dependencies/kube-object-menu-items/kube-object-menu-registry.injectable";
 import { renderFor, DiRender } from "../test-utils/renderFor";
 import type { Cluster } from "../../../main/cluster";
 import type { ApiManager } from "../../../common/k8s-api/api-manager";
-import apiManagerInjectable from "./dependencies/apiManager.injectable";
+import apiManagerInjectable from "./dependencies/api-manager.injectable";
 import { KubeObjectMenu } from "./index";
 
 describe("kube-object-menu", () => {
